requirements.txt
atlassian-python-api==3.41.1
python-dateutil
ruamel.yaml
tqdm
